# bug in groupby when key space exceeds int64 bounds